+++ src/pharmpy/modeling/__init__.py
@@ -106,7 +106,11 @@ from pharmpy.modeling.write_csv import write_csv
 
 from .allometry import add_allometry
 from .amd import run_amd, run_iiv
-from .expressions import get_observation_expression
+from .expressions import (
+    get_individual_prediction_expression,
+    get_observation_expression,
+    get_population_prediction_expression,
+)
 from .iterators import omit_data, resample_data
 from .plots import plot_individual_predictions, plot_iofv_vs_iofv
 from .units import get_unit_of
@@ -143,6 +147,7 @@ __all__ = [
     'get_doses',
     'get_doseid',
     'get_ids',
+    'get_individual_prediction_expression',
     'get_number_of_individuals',
     'get_number_of_observations',
     'get_number_of_observations_per_individual',
@@ -152,6 +157,7 @@ __all__ = [
     'get_model_covariates',
     'get_observations',
     'get_observation_expression',
+    'get_population_prediction_expression',
     'get_unit_of',
     'has_additive_error_model',
     'has_combined_error_model',