+++ tests/modeling/test_expressions.py
@@ -1,7 +1,11 @@
 import sympy
 
 from pharmpy import Model
-from pharmpy.modeling import get_observation_expression
+from pharmpy.modeling import (
+    get_individual_prediction_expression,
+    get_observation_expression,
+    get_population_prediction_expression,
+)
 
 
 def s(x):
@@ -18,3 +22,17 @@ def test_get_observation_expression(testdata):
     ) + s(
         'OPRED'
     )
+
+
+def test_get_individual_prediction_expression(testdata):
+    model = Model(testdata / 'nonmem' / 'pheno_real_linbase.mod')
+    expr = get_individual_prediction_expression(model)
+    assert expr == s('D_ETA1') * (s('ETA(1)') - s('OETA1')) + s('D_ETA2') * (
+        s('ETA(2)') - s('OETA2')
+    ) + s('OPRED')
+
+
+def test_get_population_prediction_expression(testdata):
+    model = Model(testdata / 'nonmem' / 'pheno_real_linbase.mod')
+    expr = get_population_prediction_expression(model)
+    assert expr == -s('D_ETA1') * s('OETA1') - s('D_ETA2') * s('OETA2') + s('OPRED')